Dodalem bardzo duzo min logowanie i autoryzacja, bezpieczenstwo endpointow admina, poprawiona logika uzytkownika, walidacja ownera, panel mieszkanca i logika zgloszen

+++ DominoApp/views/announ.py
@@ -9,6 +9,7 @@ from rest_framework import status
 from ..models import Announ, User
 from ..serializer import AnnounSerializer
 from DominoApp import serializer
+from DominoApp.utils import get_authenticated_user
 
 @api_view(['GET', 'POST'])
 def get_announ(request):
@@ -18,6 +19,10 @@ def get_announ(request):
         return Response(serializer.data)
     
     elif request.method == 'POST':
+        user, error_response = get_authenticated_user(request, required_role='ADMIN')
+        if error_response:
+            return error_response
+        
         serializer = AnnounSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -36,6 +41,10 @@ def announ_detail(request, pk):
         return Response(serializer.data)
     
     elif request.method == 'PUT':
+        user, error_response = get_authenticated_user(request, required_role='ADMIN')
+        if error_response:
+            return error_response
+        
         serializer = AnnounSerializer(announ, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -43,6 +52,10 @@ def announ_detail(request, pk):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     
     elif request.method == 'DELETE':
+        user, error_response = get_authenticated_user(request, required_role='ADMIN')
+        if error_response:
+            return error_response
+        
         announ.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
